+++ semana_04/enunciado.py
@@ -1,5 +1,60 @@
 import streamlit as st
 
-def ejecutar_enunciado():
-  st.subheader("Enunciado")
-  st.write("Este es el enunciado")
+def ejecutar_enunciado1():
+  st.subheader("Problema 1")
+  st.write("Implementar un programa que permita ingresar dos números y que por medio de un método recursivo calcule la división entera por restas sucesivas.")
+
+def ejecutar_enunciado2():
+  st.subheader("Problema 2")
+  st.write("Realizar un programa que permita ingresar un número y que por medio de un método recursivo invierta la posición de sus cifras (ejemplo: 2365 lo invierte a 5632).")
+
+def ejecutar_enunciado3():
+  st.subheader("Problema 3")
+  st.write("Realizar un programa que sume los dígitos de un número utilizando un algoritmo recursivo. El número se debe ingresar por teclado.")
+
+def ejecutar_enunciado4():
+  st.subheader("Problema 4")
+  st.write("Realizar un algoritmo recursivo para multiplicar dos números.")
+
+def ejecutar_enunciado5():
+  st.subheader("Problema 5")
+  st.write("Elabore un programa que por medio de un método recursivo sume los elementos de un vector. ")
+
+def ejecutar_enunciado6():
+  st.subheader("Problema 6")
+  st.write("Realice un algoritmo recursivo para determinar si un número ingresado por teclado es positivo o negativo.")
+
+def ejecutar_enunciado7():
+  st.subheader("Problema 7")
+  st.write("Calcular la suma de dígitos de un número “n” con un algoritmo recursivo.")
+
+def ejecutar_enunciado8():
+  st.subheader("Problema 8")
+  st.write("Convertir un número entero positivo a binario, con un algoritmo recursivo.")
+
+def ejecutar_enunciado9():
+  st.subheader("Problema 9")
+  st.write("Analice el siguiente código y responda. ¿Qué hace el siguiente código?")
+  code = '''
+    Este algoritmo multiplica dos números a y b
+Donde:
+	Si b es par:
+	mimetodo(2*a,b/2) equivale a:
+		2a * b/2 = ab
+	Si b es impar:
+	mimetodo(2*a,b/2) + a es igual a:
+		mimetodo(2*a,(b-1)/2) + a
+	Equivale a:
+		2a * (b-1)/2 + a = ab
+    '''
+  st.code(code, language='java')
+
+  st.image("enunciado9.png", caption="Enunciado 9")
+
+def ejecutar_enunciado10():
+  st.subheader("Problema 10")
+  st.write("Calcular la potencia n x con un algoritmo recursivo.")
+
+def ejecutar_enunciado11():
+  st.subheader("Problema 11")
+  st.write("Calcular el Máximo Común Divisor de 2 números con un algoritmo recursivo.")
