+++ semana_04/semana_04.py
@@ -5,8 +5,59 @@ from .analisis import *
 from .graficas import *
 
 def ejecutar_semana_04():
-  st.header(f"Semana 04")
-  ejecutar_enunciado()
-  ejecutar_implementacion()
-  ejecutar_analisis()
-  ejecutar_graficas()
+  st.header(f"Semana 03")
+
+  # Ejercicio 1
+  ejecutar_enunciado1()
+  ejecutar_implementacion1()
+  ejecutar_analisis1()
+
+  # Ejercicio 2
+  ejecutar_enunciado2()
+  ejecutar_implementacion2()
+  ejecutar_analisis2()
+
+  # Ejercicio 3
+  ejecutar_enunciado3()
+  ejecutar_implementacion3()
+  ejecutar_analisis3()
+
+  # Ejercicio 4
+  ejecutar_enunciado4()
+  ejecutar_implementacion4()
+  ejecutar_analisis4()
+
+  # Ejercicio 5
+  ejecutar_enunciado5()
+  ejecutar_implementacion5()
+  ejecutar_analisis5()
+
+  # Ejercicio 6
+  ejecutar_enunciado6()
+  ejecutar_implementacion6()
+  ejecutar_analisis6()
+
+  # Ejercicio 7
+  ejecutar_enunciado7()
+  ejecutar_implementacion7()
+  ejecutar_analisis7()
+
+  # Ejercicio 8
+  ejecutar_enunciado8()
+  ejecutar_implementacion8()
+  ejecutar_analisis8()
+
+  # Ejercicio 9
+  ejecutar_enunciado9()
+  ejecutar_implementacion9()
+  ejecutar_analisis9()
+
+   # Ejercicio 10
+  ejecutar_enunciado10()
+  ejecutar_implementacion10()
+  ejecutar_analisis10()
+
+   # Ejercicio 11
+  ejecutar_enunciado11()
+  ejecutar_implementacion11()
+  ejecutar_analisis11()
